Fixing HitRecord and Shape order
