+++ myRayTracing/src/myRayTracing.jl
@@ -26,14 +26,14 @@ export Camera, fire_single_ray, OrthogonalCamera, PerspectiveCamera, aperture_de
 include("ImageTracer.jl")
 export ImageTracer, fire_ray, fire_all_rays!
 
-include("hitrecord.jl")
-export HitRecord, is_close
+# include("hitrecord.jl")
+# export HitRecord, is_close
 
 include("materials.jl")
 export Pigment, UniformPigment, CheckeredPigment, ImagePigment, get_color, BRDF, DiffuseBRDF, eval, Material
 
 include("shapes.jl")
-export Shape, Sphere, Plane, ray_intersection, quick_ray_intersection, union_shape, intersec_shape, diff_shape
+export Shape, Sphere, Plane, ray_intersection, quick_ray_intersection, HitRecord, is_close, union_shape, intersec_shape, diff_shape
 
 include("world.jl")
 export World, add_shape!, get_shapes, get_single_shape, ray_intersection, is_point_visible
